SE router

+++ trains/singleTask/model/router.py
@@ -26,9 +26,9 @@ class router(nn.Module):
         super(router, self).__init__()
         self.avg_pool = nn.AdaptiveAvgPool2d(1)
         self.fc = nn.Sequential(
-            nn.Linear(channel, channel * ratio, bias=False),
+            nn.Linear(channel, int(channel * ratio), bias=False),
             nn.ReLU(inplace=True),
-            nn.Linear(channel * ratio, channel, bias=False),
+            nn.Linear(int(channel * ratio), channel, bias=False),
             nn.Sigmoid()
         )
         self.out = nn.Linear(channel, dim_num, bias=False)
